Move `SubspaceCodec` into separate module, but re-export it at the root as before
