+++ crates/subspace-solving/tests/integration/codec.rs
@@ -1,0 +1,3 @@
+mod cpu;
+#[cfg(feature = "cuda")]
+mod cuda;
