+++ crates/subspace-solving/tests/integration/main.rs
@@ -1,3 +1,1 @@
-mod cpu;
-#[cfg(feature = "cuda")]
-mod cuda;
+mod codec;
